feat: refactor workout section handling and implement caching in workout_controller
